requirements.txt
streamlit
requests
pypdf
google-generativeai
python-dotenv